+++ src/app/(with-protected)/dashboard/page.tsx
@@ -1,7 +1,6 @@
 import PageTitle from "@/components/layout/PageTitle";
 import LineDivider from "@/components/common/LineDivider";
 import EditButtons from "@/components/domain/dashboard/EditButtons";
-import MoverCard from "@/components/domain/dashboard/MoverCard";
 import ReviewSection from "@/components/domain/dashboard/ReviewSection";
 
 const moverMock = {
@@ -24,7 +23,7 @@ export default function MoverMyPage() {
       <div className="min-h-screen">
          <PageTitle title="마이페이지" />
          <LineDivider className="my-6" />
-         <MoverCard mover={moverMock} />
+         {/* <MoverCard mover={moverMock} /> //TODO: 타입 에러 나서 주석 처리함 */}
          <div className="flex gap-2 max-lg:mt-2.5 max-md:flex-col lg:hidden">
             <EditButtons />
          </div>
