+++ src/lib/api/auth/requests/profile.ts
@@ -1,8 +1,5 @@
 import { MoverBasicInfoInput } from "@/lib/schemas/dashboard.schema";
-import {
-   MoverProfileInput,
-   MoverProfileRequestInput,
-} from "@/lib/schemas/profile.schema";
+import { MoverProfileRequestInput } from "@/lib/schemas/profile.schema";
 import { tokenFetch } from "@/lib/utils";
 
 //기사님 기본정보 수정 api
@@ -10,6 +7,18 @@ async function updateMoverInfo(data: MoverBasicInfoInput) {
    const url = "/dashboard/edit/mover";
    return await tokenFetch(url, {
       method: "PATCH",
+      headers: {
+         "Content-Type": "application/json",
+      },
+      body: JSON.stringify(data),
+   });
+}
+
+//기사님 프로필 등록 api
+async function createProfile(data: MoverProfileRequestInput) {
+   const url = "/profile/mover";
+   return await tokenFetch(url, {
+      method: "POST",
       headers: {
          "Content-Type": "application/json",
       },
@@ -29,4 +38,4 @@ async function updateMoverProfile(data: MoverProfileRequestInput) {
    });
 }
 
-export { updateMoverInfo, updateMoverProfile };
+export { updateMoverInfo, createProfile, updateMoverProfile };
